[EI-127] [web] refactor: Refactor dialog-menu component using compound pattern
