+++ web/app/ui/components/view/molocule/dialog-menu/index.ts
@@ -1,0 +1,8 @@
+import DialogMenuRoot from './dialog-menu-root';
+import DialogMenuItem from './dialog-menu-item';
+
+const DialogMenu = Object.assign(DialogMenuRoot, {
+  Item: DialogMenuItem,
+});
+
+export default DialogMenu;
